+++ header.php
@@ -19,6 +19,45 @@
       <title><?php wp_title(''); ?> (<?php bloginfo('name'); ?>)</title>
       <?php } ?>
       <link href="https://s3.amazonaws.com/wrdsb-ui-assets/public/master.css" rel="stylesheet" media="all" />
+  
+      <?php if ( function_exists('max_mega_menu_is_enabled') && max_mega_menu_is_enabled('top') ) : ?>
+      <style type="text/css">
+      <?php if (wrdsb_i_am_a_corporate_site()):?>
+        .my-navbar{
+          margin-bottom:0px;
+        }
+      <?endif;?>
+      #navbar-header{
+        background-color: <?php echo getNavbarBackgroundColor() ?>;
+        height:80px !important;
+        background-position:4px 15px !important;
+      }
+
+      #navbar-header a.navbar-brand {
+        margin-top:20px !important;
+      }
+
+      .togglesearch {
+        top:17px;
+        right: 50px;
+      }
+      @media (max-width:768px){
+      .mobileMenu{
+        float: right;
+        z-index:10;
+        display:inline-block;
+      }
+      }
+
+      .navbar-search input[type="text"] {
+        width:100%;
+        }
+      
+        #mobileSearch{
+        border-radius: 5px;
+        }
+      </style>
+      <?php endif;?>
        <!-- icons -->
       <link href="https://s3.amazonaws.com/wrdsb-ui-assets/<?php echo $GLOBALS['wrdsbvars']['asset_version']; ?>/images/icon-60x60.png" rel="apple-touch-icon" />
       <link href="https://s3.amazonaws.com/wrdsb-ui-assets/<?php echo $GLOBALS['wrdsbvars']['asset_version']; ?>/images/icon-76x76.png" rel="apple-touch-icon" sizes="76x76" />
@@ -127,9 +166,10 @@
       <div id="fb-root"></div>
       <script async defer crossorigin="anonymous" src="https://connect.facebook.net/en_GB/sdk.js#xfbml=1&version=v4.0"></script>
       <? } ?>
- 
+      
       <!-- header -->
       <div id="masthead">
+      <?php if (!wrdsb_i_am_a_corporate_site() || (wrdsb_i_am_a_corporate_site() && !function_exists('max_mega_menu_is_enabled'))): ?>
         <div class="container-top">
           <div class="header">
             <div class="container">
@@ -168,37 +208,43 @@
           </div>
         </div>
       </div>
-      <div class="navbar my-navbar" role="navigation" aria-labelledby="navbar-header">
-        <div id="navbar-header">
-          <button type="button" class="navbar-toggle togglesearch" data-toggle="collapse" data-target=".navbar-search">
-          <span class="sr-only">Toggle navigation</span>
-          <span class="icon-search"><img src="https://wrdsb-ui-assets.s3.amazonaws.com/public/2/2.0.0/images/search.gif" style="width: 25px; height: 25px;" /></span>
-          </button>
-          <button type="button" class="navbar-toggle togglenav" data-toggle="collapse" data-target=".navbar-collapse">
-          <span class="sr-only">Toggle navigation</span>
-          <span class="icon-bar"></span>
-          <span class="icon-bar"></span>
-          <span class="icon-bar"></span>
-          </button>
-          <?php $strl = get_bloginfo('name');
-          $tempstrl = explode(" ", $strl);
-          $newstrl = '';
-          foreach($tempstrl as $a) {
-          if ($a=='District' or $a=='Public' or $a=='Collegiate' or $a=='Secondary' or $a=='High') {
-          $newstrl.= "<br />";
-          }
-          $newstrl.=" ".$a;
-          }
-          ?>
-          <a class="navbar-brand" href="<?php echo home_url(); ?>/"><?php echo $newstrl; ?></a>
-        </div>
+      <?php endif; ?>
+
+<!-- ---------------------------MMM DISABLED NAVBAR----------------------------------------------------------------- -->
+      <?php if (!function_exists('max_mega_menu_is_enabled') || !max_mega_menu_is_enabled('top') ) : ?>
+        <div class="navbar my-navbar" role="navigation" aria-labelledby="navbar-header">
+        
+          <div id="navbar-header">
+            <button type="button" class="navbar-toggle togglesearch" data-toggle="collapse" data-target=".navbar-search">
+            <span class="sr-only">Toggle navigation</span>
+            <span class="icon-search"><img src="https://wrdsb-ui-assets.s3.amazonaws.com/public/2/2.0.0/images/search.gif" style="width: 25px; height: 25px;" /></span>
+            </button>
+            <button type="button" class="navbar-toggle togglenav" data-toggle="collapse" data-target=".navbar-collapse">
+            <span class="sr-only">Toggle navigation</span>
+            <span class="icon-bar"></span>
+            <span class="icon-bar"></span>
+            <span class="icon-bar"></span>
+            </button>
+            
+            <?php $strl = get_bloginfo('name');
+            $tempstrl = explode(" ", $strl);
+            $newstrl = '';
+            foreach($tempstrl as $a) {
+            if ($a=='District' or $a=='Public' or $a=='Collegiate' or $a=='Secondary' or $a=='High') {
+            $newstrl.= "<br />";
+            }
+            $newstrl.=" ".$a;
+            }
+            ?>
+            <a class="navbar-brand" href="<?php echo home_url(); ?>/"><?php echo $newstrl; ?></a>
+          </div>
         
         <div class="collapse navbar-search" role="search" aria-labelledby="mobileSearch">
           <form action="<?php echo home_url(); ?>/" method="get">
             <input aria-label="Search" type="text" name="s" id="mobileSearch" value="<?php the_search_query(); ?>" placeholder="Search" />
           </form>
         </div>
-        <div id="menu" class="container" role="navigation" aria_label="Menu">
+        <div id="menu" role="navigation" aria_label="Menu">
           <?php if (has_nav_menu('top')) {
           wp_nav_menu(array('theme_location' => 'top', 'menu_class' => 'nav nav-justified', 'container_class' => 'collapse navbar-collapse'));
           } else {
@@ -206,13 +252,59 @@
           } ?>
         </div>
         </div><!-- /.navbar -->
-        </div><!-- /.container-top -->
+        <?php endif; ?>
+        </div><!-- /#masthead -->
+<!-- ---------------------------MMM ENABLED NAVBAR----------------------------------------------------------------- -->
+
+        <?php if (function_exists('max_mega_menu_is_enabled') && max_mega_menu_is_enabled('top') ) : ?>
+        <div class="navbar my-navbar" role="navigation" aria-labelledby="navbar-header">
+        
+          <div id="navbar-header">
+            <button type="button" class="navbar-toggle togglesearch" data-toggle="collapse" data-target=".navbar-search">
+            <span class="sr-only">Toggle navigation</span>
+            <span class="icon-search"><img src="https://wrdsb-ui-assets.s3.amazonaws.com/public/2/2.0.0/images/search.gif" style="width: 25px; height: 25px;" /></span>
+            </button>
+            
+            <div id="menu" role="navigation" class="mobileMenu" aria_label="Menu">
+                <?php if (has_nav_menu('top')) {
+                wp_nav_menu(array('theme_location' => 'top', 'menu_class' => 'nav nav-justified', 'container_class' => 'collapse navbar-collapse'));
+                } else {
+                wp_page_menu(array('depth' => 1, 'show_home' => true, 'menu_class' => 'collapse navbar-collapse' ));
+                } ?>
+            </div>
+            <?php $strl = get_bloginfo('name');
+            $tempstrl = explode(" ", $strl);
+            $newstrl = '';
+            foreach($tempstrl as $a) {
+            if ($a=='District' or $a=='Public' or $a=='Collegiate' or $a=='Secondary' or $a=='High') {
+            $newstrl.= "<br />";
+            }
+            $newstrl.=" ".$a;
+            }
+            ?>
+            <a class="navbar-brand" href="<?php echo home_url(); ?>/"><?php echo $newstrl; ?></a>
+          </div>
+        
+        
+        <div class="collapse navbar-search" role="search" aria-labelledby="mobileSearch">
+          <form action="<?php echo home_url(); ?>/" method="get">
+            <input aria-label="Search" type="text" name="s" id="mobileSearch" value="<?php the_search_query(); ?>" placeholder="Search" />
+          </form>
+        </div>
+        
+        </div><!-- /.navbar -->
+        <?php endif; ?>
+
+<!-- ----------------------------------------MEGAMENU END-------------------------------------------------------------- -->
+
+
 
         <?php if (is_home() && is_front_page() && wrdsb_i_am_a_corporate_site()) {
           echo do_shortcode('[slick-slider design="prodesign-28" read_more_text="Learn more" link_target="blank" sliderheight="300" dots_design="design-6" autoplay_interval="5000" speed="900" fade="true" focus_pause="true" category="www"]');
         } ?>
 
-        <?php if (is_front_page() && !wrdsb_i_am_a_corporate_site()) {
+
+        <?php if (is_front_page()) {
 
         // if we have an alert
         if (function_exists('stswr_alerts_get_current_alert') && stswr_alerts_get_current_alert('id') !== '0') {
@@ -235,3 +327,4 @@
         <?php if (!is_front_page()) { ?>
         <?php the_breadcrumb(); ?>
         <?php } ?>
+